move tests, small fix to modify filters
